+++ Code/GreenOnion/util/functions.hpp
@@ -2,7 +2,7 @@
 #include <string>       // for string
 #include <vector>       // for vector
 
-void pad(std::string &inputString, int minimumLength, char paddingChar);
+void pad(std::string &inputString, size_t minimumLength, char paddingChar);
 void sleep(int milliseconds);
 void split(std::vector<std::string> &result, std::string str, char delim);
 std::string key_from_exponent_and_base_packet(std::string basePacket, std::string exponent);
